converted numpy.float64 values that may conflict with orjson.

+++ review.py
@@ -6,7 +6,7 @@ max_battery_inverter_power = max_inverter_limit * len(battery_inverter_serials)
 
 
 def _energy_power(power: pd.Series, utc: pd.Series) -> float:
-    return round(np.trapz(y=power.fillna(0) / 1e3, x=utc / 3600), 3)
+    return float(round(np.trapz(y=power.fillna(0) / 1e3, x=utc / 3600), 3))
 
 
 def _plot_list(power: pd.Series) -> dict[str, list]:
@@ -27,7 +27,7 @@ def create_day_review(df: pd.DataFrame) -> dict:
     response["missing_energy"] = _energy_power(
         np.maximum(df["consumer_power"] - df["producer_power"], 0), df.index
     )
-    response["plot_limit"] = max(200, df["producer_power"].max())
+    response["plot_limit"] = float(max(200, df["producer_power"].max()))
     if max_battery_inverter_power > 0:
         response["max_battery_inverter_power"] = max_battery_inverter_power
         response["battery_unreachable_energy"] = _energy_power(
